+++ packages/react-magma-dom/jest.config.js
@@ -1,11 +1,9 @@
-const baseConfig = require("../../jest.config")
-const name = __dirname.split('/').pop() 
+const baseConfig = require('../../jest.config');
+const name = __dirname.split('/').pop();
 
 module.exports = {
   ...baseConfig,
-  collectCoverageFrom: [
-    `**/packages/${name}/src/**/*.{js,jsx,ts,tsx}`,
-  ],
+  collectCoverageFrom: [`**/packages/${name}/src/**/*.{js,jsx,ts,tsx}`],
   displayName: {
     name,
     color: 'magenta',
@@ -13,4 +11,4 @@ module.exports = {
   projects: ['.'],
   rootDir: '../..',
   testMatch: [`**/${name}/**/?(*.)+(spec|test).{js,ts,mjs}`],
-}
+};