+++ packages/react-magma-dom/src/components/LoadingIndicator/LoadingIndicator.test.js
@@ -11,7 +11,10 @@ describe('Loading Indicator', () => {
     jest.useFakeTimers();
   });
 
-  afterEach(() => {
+  afterEach(async () => {
+    await act(async () => {
+      jest.runAllTimers();
+    });
     jest.useRealTimers();
   });