+++ packages/react-magma-dom/src/components/Select/MultiSelect.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
+import { render, fireEvent, act } from '@testing-library/react';
 import { Select as MultiSelect } from '.';
 
 describe('Select', () => {
@@ -82,6 +82,8 @@ describe('Select', () => {
   });
 
   it('should not select an item when typing and select is closed', () => {
+    // Use fake timers here for downshift's debounce on input change.
+    jest.useFakeTimers();
     const labelText = 'Label';
     const items = ['Red', 'Blue', 'Green'];
     const { getByLabelText, queryByText } = render(
@@ -94,9 +96,13 @@ describe('Select', () => {
 
     fireEvent.keyDown(renderedSelect, { key: 'r' });
 
+    act(() => jest.runAllTimers());
+
     expect(
       queryByText(items[0], { selector: 'button' })
     ).not.toBeInTheDocument();
+
+    jest.useRealTimers();
   });
 
   it('should allow for selection of multiple items', () => {
@@ -355,6 +361,8 @@ describe('Select', () => {
   });
 
   it('should not open select when clicking another key other than the enter or spacebar', () => {
+    // Use fake timers here for downshift's debounce on input change.
+    jest.useFakeTimers();
     const labelText = 'Label';
     const items = ['Red', 'Blue', 'Green'];
     const { getByLabelText, queryByText } = render(
@@ -369,7 +377,11 @@ describe('Select', () => {
       key: 'a',
     });
 
+    act(() => jest.runAllTimers());
+
     expect(queryByText(items[0])).not.toBeInTheDocument();
+
+    jest.useRealTimers();
   });
 
   it('should show an error message', () => {