+++ packages/react-magma-dom/src/components/Tabs/Tab.tsx
@@ -1,25 +1,20 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import { ThemeContext } from '../../theme/ThemeContext';
-import { css, jsx } from '@emotion/core';
+import { css } from '@emotion/core';
 import isPropValid from '@emotion/is-prop-valid';
-import {
-  TabsIconPosition,
-  TabsOrientation,
-  TabsBorderPosition,
-  TabsContext,
-} from '.';
-import { Omit, useForceUpdate, useForkedRef, XOR } from '../../utils';
+import { TabsIconPosition, TabsBorderPosition, TabsContext } from '.';
+import { TabsOrientation } from './shared';
+import { useForceUpdate, useForkedRef } from '../../utils';
 import { TabsContainerContext } from './TabsContainer';
 import { ThemeInterface } from '../../theme/magma';
 
-export interface BaseTabProps
-  extends Omit<React.ButtonHTMLAttributes<HTMLButtonElement>, 'children'> {
-  disabled?: boolean;
+export interface TabProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   icon?: React.ReactElement<any> | React.ReactElement<any>[];
   iconPosition?: TabsIconPosition;
-  index?: number;
   isActive?: boolean;
+  isDisabled?: boolean;
   isFullWidth?: boolean;
   isInverse?: boolean;
   orientation?: TabsOrientation;
@@ -28,17 +23,9 @@ export interface BaseTabProps
   theme?: any;
 }
 
-interface TabChildrenProps extends BaseTabProps {
-  children: JSX.Element | string;
-}
-
-interface TabComponentProps extends BaseTabProps {
-  component: React.ReactNode;
-}
-
-export type TabProps = XOR<TabChildrenProps, TabComponentProps>;
-
-const StyledTabsChild = styled('li', { shouldForwardProp: isPropValid })<{
+export const StyledTabsChild = styled('li', {
+  shouldForwardProp: isPropValid,
+})<{
   borderPosition?: TabsBorderPosition;
   isActive?: boolean;
   isFullWidth?: boolean;
@@ -105,7 +92,7 @@ function getFlexDirection(position: TabsIconPosition) {
   }
 }
 
-const TabStyles = props => css`
+export const TabStyles = props => css`
   align-items: center;
   background: transparent;
   border: 0;
@@ -124,7 +111,11 @@ const TabStyles = props => css`
   line-height: ${props.theme.typeScale.size02.lineHeight};
   height: 100%;
   justify-content: ${props.iconPosition === 'left' ? 'flex-start' : 'center'};
-  opacity: ${props.disabled ? 0.4 : props.isInverse ? 0.7 : 1};
+  opacity: ${props.disabled
+    ? 0.4
+    : props.isInverse && !props.isActive
+    ? 0.7
+    : 1};
   padding: 13px 20px;
   position: relative;
   pointer-events: ${props.disabled ? 'none' : ''};
@@ -153,11 +144,12 @@ const TabStyles = props => css`
       ? props.theme.colors.shade02
       : props.theme.colors.shade01};
     color: ${props.isActive
-      ? ''
+      ? props.isInverse
+        ? props.theme.colors.neutral08
+        : props.theme.colors.primary
       : props.isInverse
       ? props.theme.colors.neutral08
       : props.theme.colors.neutral02};
-    opacity: ${props.disabled ? 0.4 : 1};
   }
 
   &:focus {
@@ -170,46 +162,10 @@ const TabStyles = props => css`
 `;
 
 const StyledTab = styled('button', { shouldForwardProp: isPropValid })<
-  BaseTabProps
+  TabProps
 >`
   ${TabStyles}
 `;
-
-export const StyledCustomTab: React.FunctionComponent<TabComponentProps> = ({
-  children,
-  component,
-  icon,
-  index,
-  style,
-  onClick,
-  ref,
-  ...props
-}) => {
-  if (React.isValidElement(component) && React.isValidElement(component)) {
-    const cloneElement = (element, newProps) => {
-      return jsx(element.type, {
-        key: element.key,
-        ref: element.ref,
-        ...element.props,
-        ...newProps,
-      });
-    };
-
-    return cloneElement(component, {
-      ...props,
-      css: TabStyles(props),
-      ...style,
-      onClick,
-      ref,
-      children: (
-        <>
-          {icon}
-          {component.props.children}
-        </>
-      ),
-    });
-  }
-};
 
 function getIconMargin(props) {
   if (props.isIconOnly) {
@@ -230,7 +186,7 @@ function getIconMargin(props) {
   }
 }
 
-const StyledIcon = styled.span<{
+export const StyledIcon = styled.span<{
   iconPosition: TabsIconPosition;
   isIconOnly?: boolean;
 }>`
@@ -243,22 +199,9 @@ const StyledIcon = styled.span<{
   }
 `;
 
-function instanceOfComponentTab(object: any): object is TabComponentProps {
-  return 'component' in object && !('children' in object);
-}
-
-function instanceOfChildrenTab(object: any): object is TabChildrenProps {
-  return !('component' in object) && 'children' in object;
-}
-
 export const Tab: React.FunctionComponent<TabProps> = React.forwardRef(
-  (
-    props: Omit<React.PropsWithChildren<TabProps>, 'children'>,
-    forwardedRef: React.Ref<any>
-  ) => {
-    let component;
-    let children;
-    const { icon, testId, ...rest } = props;
+  (props: TabProps, forwardedRef: React.Ref<any>) => {
+    const { children, icon, isDisabled, testId, ...rest } = props;
     const { activeTabIndex } = React.useContext(TabsContainerContext);
     const { buttonRefArray, registerTabButton } = React.useContext(TabsContext);
     const ownRef = React.useRef<HTMLDivElement>();
@@ -289,12 +232,6 @@ export const Tab: React.FunctionComponent<TabProps> = React.forwardRef(
       forceUpdate();
     }, []);
 
-    if (instanceOfComponentTab(props)) {
-      component = props.component;
-    } else if (instanceOfChildrenTab(props)) {
-      children = props.children;
-    }
-
     const theme = React.useContext(ThemeContext);
     const isIconOnly = !children;
 
@@ -306,58 +243,37 @@ export const Tab: React.FunctionComponent<TabProps> = React.forwardRef(
 
     return (
       <StyledTabsChild
-        aria-selected={isActive}
         borderPosition={borderPosition}
         data-testid="tabContainer"
         isActive={isActive}
         isFullWidth={isFullWidth}
         isInverse={isInverse}
-        ref={ref}
         orientation={orientation}
-        onClick={e => changeHandler(index, e)}
-        role="tab"
         theme={theme}
       >
-        {component ? (
-          <StyledCustomTab
-            {...rest}
-            component={component}
-            data-testid={testId}
-            iconPosition={tabIconPosition}
-            icon={
-              icon && (
-                <StyledIcon iconPosition={tabIconPosition}>{icon}</StyledIcon>
-              )
-            }
-            index={index}
-            isActive={isActive}
-            isInverse={isInverse}
-            isFullWidth={isFullWidth}
-            orientation={orientation}
-            theme={theme}
-          />
-        ) : (
-          <StyledTab
-            {...rest}
-            data-testid={testId}
-            iconPosition={tabIconPosition}
-            isActive={isActive}
-            isInverse={isInverse}
-            isFullWidth={isFullWidth}
-            orientation={orientation}
-            theme={theme}
-          >
-            {icon && (
-              <StyledIcon
-                iconPosition={tabIconPosition}
-                isIconOnly={isIconOnly}
-              >
-                {icon}
-              </StyledIcon>
-            )}
-            {children}
-          </StyledTab>
-        )}
+        <StyledTab
+          {...rest}
+          aria-selected={isActive}
+          data-testid={testId}
+          disabled={isDisabled}
+          iconPosition={tabIconPosition}
+          isActive={isActive}
+          isInverse={isInverse}
+          isFullWidth={isFullWidth}
+          onClick={e => changeHandler(index, e)}
+          orientation={orientation}
+          ref={ref}
+          role="tab"
+          tabIndex={isActive ? 0 : -1}
+          theme={theme}
+        >
+          {icon && (
+            <StyledIcon iconPosition={tabIconPosition} isIconOnly={isIconOnly}>
+              {icon}
+            </StyledIcon>
+          )}
+          {children}
+        </StyledTab>
       </StyledTabsChild>
     );
   }