+++ packages/react-magma-dom/src/components/Toast/Toast.test.js
@@ -8,7 +8,10 @@ describe('Toast', () => {
     jest.useFakeTimers();
   });
 
-  afterEach(() => {
+  afterEach(async () => {
+    await act(async () => {
+      jest.runAllTimers();
+    });
     jest.useRealTimers();
   });